refactor utisl
