+++ main/utils/appUpdater.ts
@@ -2,7 +2,7 @@ import { BrowserWindow, dialog, Tray } from 'electron';
 import { autoUpdater } from 'electron-updater';
 import log from 'electron-log';
 import { setIsAppQuitting, setIsUpdateAvailable } from '..';
-import { buildContextMenu } from './menuUtil';
+import { buildContextMenu } from './menu';
 
 class AppUpdater {
     constructor() {
