+++ main/utils/createWindow.ts
@@ -2,7 +2,6 @@ import { app, BrowserWindow, nativeImage } from 'electron';
 import * as path from 'path';
 import { isDev } from './common';
 import { isAppQuitting } from '..';
-import { addAllowOriginHeader } from './cors';
 import { PROD_HOST_URL } from '../config';
 
 export function createWindow(): BrowserWindow {
@@ -29,8 +28,6 @@ export function createWindow(): BrowserWindow {
         transparent: true,
     });
     splash.maximize();
-
-    addAllowOriginHeader(mainWindow);
 
     if (isDev) {
         splash.loadFile(`../build/splash.html`);
